Add prelim versions make-nn-strus, make-acme-nn-strus, renamed other fns to make-*
Renamed functions such as node-vec to make-node-vec to avoid inadvertent
clash with variable names.

Defined preliminary versions of make-nn-stru, which constructs a
Clojure map containing a vector of node info, a Clojure map from that
info to vector indexes, and an all-zero core.matrix weight matrix.
This must be filled in later.  This function can be used both for
the analogy network and the proposition network.

Defined preliminary version of make-acme-nn-stru, which calls functions
that match two sets of propositions, and then passes the result to
make-nn-stru.  (Still need to use the result of matching the propns
in order to fill the weight matrix.)