Add comment on list-propn-families, del flatten-propn-families.
New notes file acme.nt5 clarifies that what list-propn-families
is returning is not optimal--it includes propn families in argument
lists, which is not needed since links down to propn args don't
to past the propn arg into its own components.  This means
also that flatten-propn-families won't be needed.  (It's a trivial
function to recreate, if so.)