Clean up/rename match-components-of-propn-pair, match-args.
match-components-of-propn-pair was adding mapnode ids to mapnode maps
after creating them.  Now match-args is revised and rewritten as
mapnode-map, and it adds the mapnode id.  This is used now everywhere in
match-components-of-propn-pair to create the mapnode maps.  Also changed
to match-components-of-propn-pair to return conses rather than lazyseqs.
Doesn't matter--all that matters is that a seq be returned--but I
prefered to use cons twice rather than list and then concat.