Fix add-families-wts-to-mat!, add it into make-acme-nn-strus.
Added call to add-families-wts-to-mat! to set some weights in the matrix
created in make-acme-nn-strus.  Also discovered bug in add-...-mat! :
Outer loop in it was a call to map, whose value wasn't returned.  So
the loop never ran, and so the matrix wasn't updated.  Now uses doseq.
(There's a gotcha.)

Note add-families-wts-to-mat! hasn't been properly tested, and I think
it's only adding weights in one direction, which isn't right.