Rearranged nn-stru pprint functions, added dot-print-nn-stru.
Renamed pprint-matrix-with-labels and pprint-nn-stru to
        format-matrix-with-labels and format-nn-stru,
deleting the print call in the former.  Thus these functions
return strings, which will look nice if given to print.

Then defined a new pprint-nn-stru, which just wraps print
around format-nn-stru.

Then also defined dotprint-nn-stru, which does the same thing
but replaces instances of "0.0 " with " .  " .  Note that this
assumes that values can't have more than one integer digit.