Added comments, docstrings, toward fixing weight order problem.
Link weights are not coming out right, I believe, because of mismatch
in how different elements are ordered at the time that weights are added
by the call to add-families-to-wts-to-mat! in make-acme-nn-stru.
I've verified that the index order of :indexes is the same as the
sequence order of :nodes in an nn-stru.  The problem seems to be
that add-families-wts-to-mat! processes map-nodes in an order
that has to do with pairs, which is not necessarily the order of the
nodes.  Hmm then again, it does use indexes to look up indexes.
Have to investigate further.