Done converting add-*-wts-to-mat! to common code.
add-pos-wts-to-mat! and neg-pos-wts-to-mat! now both call
add-wts-to-mat!, and a new function matched-idx-fams generates families
of indexes from families of mapnodes via LOT relationships.  This is
then passed into add-pos-wts-to-mat!, rather than the same work being
done inside it, as before.

Also fixed bug in identity-if-zero (used to be identity-if-old-zero, I
think): cases in the 'if' were swapped.

Also added 'get' to retrieval of matrix from nn-stru in format-nn-stru,
for the sake of clarity.

Still need to clean up the confusing cluster of code in
make-acme-nn-stru, which might involve revising some of the functions
mentioned above.