Add temp fns rec-sym?, comp-sym? for testing
comp-sym? is the old symmetric?, i.e. same as what's in my versions of
core.matrix and vectorz-clj at the moment.  Applies test to
a elements of a lazy seq via for.

rec-sym? is a new version of the same test using recur, and avoiding
creating a list comprehension.

I want to do speed comparisons.