Removed method field :poss-mapnode? from nn-core
poss-mapnode? (which should have been called 'poss-node?') did *exactly*
the same thing as the map in :id-to-idx.  It was no more than a function
wrapper around the latter--which, after all, can be treated as a
function in Clojure.  The intent of poss-mapnode? was to create a
predicate that said whether something was a possible node.  But the
map id-to-idx returns nil if an id doesn't exist there, or the index
if it does.  So it can be used as a predicate.