Fixed bug due to bad analogy-mask in make-person.
The analogy-mask create code was just a placeholder--not intended
for use.  but then I decided to make both of the masks into
all zeros for now, which propn-mask wasn't supposed to be but was.
So now there has to be code added (from communic.clj eg) that will
populate the masks with 1's.  But they start as zero-vectors.