Fixed bugs in make-person.
Fixed a remaining badly sized def of analogy mask--I thought
I fixed it last time, but didn't.  And fixed the bug that
was causing an NPE when filling the masks.  The problem was
that I was using raw propns rather than their ids.  Also switched
from domap to doseq for the iteration through propns for each
of these tasks.