Added test for no propn before receiving propn, moved mask tests
Moved node-unmasked? and unmask! to nets.clj.
Defined node-masked?
(why bother to negate node-unmasked? when it's just a test for 1 vs 0?)
Changed uses of these fns in communic.clj and person.clj to reflect that.

communic.clj:

Renamed add-to-analogy-net to try-add-to-analogy-net.

Added node-still-unmasked?

In receive-propn, added test to make sure that propn hasn't already
been unmasked before calling add-to-propn-net and
try-add-to-analogy-net.  (For the sake of the latter.)