Renamed :propn-to-family-propn-idxs and related fns
Renamed :propn-to-family-propn-idxs to :propn-to-descendant-propn-idxs,
and likewise for functions that define it and use it.  This reflects
the fact that this data structure maps propn ids to indexes of descendant
propns of that propn (and of the propn itself).  i.e. for a given
propn, lists idx of it, idxs of propns that are its args, and so on
recursively.

What this data structure *doesn't* list are idxs of propns related
by extended family upward as well, so this one is about descendants,
not all extended family relations between propns.