Comments, removed :propn-mn-to-fam-idxs from analogynet
:propn-mn-to-fam-idxs wasn't being used.  However, I kept the
code for it in the file, commented out, in case it's needed later.

The other changes:
Commented out old testing versions of functions.
Removed comments that said that certain functions or data strus
needed to be tested yet, since they were used to successfully recreate
the crime3 analogy net so that it matched the one generated by popco1.