Added working matched-and-unmatched-links
Also deleted match-links, since now redundant and not as useful.
Also note rename of file/ns from popco1comppare.clj to popco1comp.clj.

Also renamed analogy.x to popco.x, to function as a central repository
for code that I've deleted but am not yet ready to consign to the
darkness of git history.