Added decay const, 1/decay in mask for special nodes
i.e. the decay constant, 0.9, in next-activns, is now defined by a
def, and in make-person, I set the mask element for the special
nodes SEMANTIC and SALIENT to 1 divided by this.  That way, when
I multiply the mask on the result of the Grossberg settling at the
end of each invocation of next-activns, it undoes the decay on the
special node.