Renamed propn/clone-propn-net to propn/clone, altered ns dec in popco.clj
Renaming clone-propn-net to clone allows a simple, unified name scheme
for cloning different popco objects, although it requires qualifying
each such use with a ns reference.  But that's OK.  However, doing this
caused a name clash with core.matrix/clone in popco.clj, because I was
`use`ing core.matrix as well as most of my popco ns'es.  So I changed
popco.clj to only require most of them, with abbreviations.  We'll
see how that goes--whether it's too annoying at the repl.