Changed propn-wts-from-analogy-activns to pass dim
In update-propn-wts-from-analogy-activns, I was passing the linger
matrix to propn-wts-from-analogy-activns, but the latter only used it to
get the dimensions in order to create a fresh square matrix.  This was
misleading and unnecessary.  Now we figure out the dimension in the
first functon, and only pass that number into the second one.