Wrapped mset! in functions in nets
Now mset! appears only in nets.clj, and there is also an emap! and add!
in update.clj.  I was doing this so that I could replace these few
defs with non-! versions, and thus be able to use persistent-vector
in order to test run it with BigDecimals.  This turns out to be more
trouble than it's worth.  When you switch from ! functions to !-less
functions, you have to start assoc'ing the results into the data
structures.  And I don't want the code to do that, in general.  So it
ends up being a silly excercise.