Ensured that make-* init fns are fully non-lazy.
I want laziness in iteration, but having laziness in other dimensions
raises the possibility of mysterious bugs due to poking around in the
data from the repl.  Added e.g. 'vec' to make sure that what's inside
persons at the beginning of the evolution of the population is fully
realized.