defined rand-idx via defmulti, commented out sfmt
SFMT19937 is not worth the trouble (at this point, anyway).
MersenneTwister is good, and SFMT19937 only provides a next int
function, where the int can range over the entire range of Integer.
Trying to convert this into an int between 0 and n can be done,
but why bother?  I don't even know whether the number is in
[Integer/MIN_VALUE, Integer/MAX_VALUE), or some other way of dealing
with the endpoints.  I want [...) for next-idx.