Moved id seqs to propns.clj, fixed bug
The bug was that I was trying to create an analogy net with only
part of the propns on one side, and then was trying to add a conceptual
link to a node that wasn't there, I think.

But that revealed what's mabye a flaw in the person creation system,
in that I'm unmasking all analogy nodes for which there are propn nodes,
but I'm not sure that's what I want when I'm trying to impose a bias.