Replaced constants +one+, +neg-one+, +zero+ with literals.
At one point I replaced 1.0, -1.0, and 0.0 with these constant variables
(along with other constants), so that I could experiment with using
BigDecimals or Ratios rather than Doubles.  I still like the idea of
not having literals anywhere in the code, but it's just harder to read
something like "cn/+one+" when it means something simple: 1.0.  Not
worth the trouble.

(Using BigDecimals and Ratios was disastrous. BigDecimals are slow, and
Ratios are disastrous: Repeatedly multiplying fractions in the matrix
multiplications quickly generates fractions with huge numerators and
denominators.  The entire system crawls, and Java eventually runs out of
memory.)