Added fns, data strus to return matrix of actual links
In propn nets, links can have zero weight, so there needs to be a
way of testing for the existence of links other than testing weights.
This isn't needed for routine operation, but you need it to report
network structures in order to display them, for example.
I added `links` function to NNMat, and defined it for PropnNet
in terms of a new record field :link-mat, and for AnalogyNet
simply by returning the regular weight matrix using `wt-mat`.