Wrote node-entry-to-node.
This is analogous to node in gexf_static, but operates on a single
entry of a node-data hashmap.  Next step is to do the same for edges.
Then map these functions over node-data and edge-data and pass
them to the top-level graph function from gexf-static (which probably
doesn't need to be modified except wrt the mode=static vs mode=dynamic).