Rewrite of conc-specs-to-idx-multiplier-pairs to avoid NPEs from non-existent mapnodes
I thought I committed this yesterday, but apparently not successfully.

conc-specs-to-idx-multiplier-pairs:
It was possible--no, easy--to pass conceptual relation specifications
(a weight and two predicate ids) to this function, which caused it to
pass back pairs with the weight and nil as the second element.  This
caused an NPE when later code tried to link from SEMANTIC to a node
indexed by nil.  Naturally.

Now conc-specs-to-idx-multiplier-pairs does not return such pairs,
stripping them out, after issuing a warning to stderr if (new variable)
constants/+warnings?+ is true.

As part of the rewrite, created conc-specs-to-idx-multiplier-pair and
swap-2nd-3rd.