fixed bug in "wait for csvs", decommissioned submitR*, cleanup
I was collecting the names of the csvs in submitRead2multirunRA.job,
but then added ability to delay until the popco job finished.
that meant that the csv names were collected before all csvs were
created.  bad idea.

so now csv names are collected in read2multirunRA.job, and I'm just
not going to use submitRead2multirunRA.sh, which just added an
unnecessary layer given that I always run this stuff from the
csv directory and write to that same directory.